+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Debbuging WIP...
+
 ### Updated
 
-Debbuging WIP...
+* UX: modified triggers (hover to load, click to show)
+* core: load dependency scripts eagerly
 
 
 ## ~~[0.2.2]~~ 2023-10-31